+++ lib/windows/swig_interface_wrap.cxx
@@ -13,6 +13,7 @@
 #define SWIGPYTHON
 #endif
 
+#define SWIG_PYTHON_THREADS
 #define SWIG_PYTHON_DIRECTOR_NO_VTABLE
 #define SWIG_PYTHON_EXTRA_NATIVE_CONTAINERS
 
@@ -4943,7 +4944,11 @@ SWIGINTERN PyObject *_wrap_delete_SwigPyIterator(PyObject *SWIGUNUSEDPARM(self),
     SWIG_exception_fail(SWIG_ArgError(res1), "in method '" "delete_SwigPyIterator" "', argument " "1"" of type '" "swig::SwigPyIterator *""'"); 
   }
   arg1 = reinterpret_cast< swig::SwigPyIterator * >(argp1);
-  delete arg1;
+  {
+    SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+    delete arg1;
+    SWIG_PYTHON_THREAD_END_ALLOW;
+  }
   resultobj = SWIG_Py_Void();
   return resultobj;
 fail:
@@ -4967,7 +4972,11 @@ SWIGINTERN PyObject *_wrap_SwigPyIterator_value(PyObject *SWIGUNUSEDPARM(self), 
   }
   arg1 = reinterpret_cast< swig::SwigPyIterator * >(argp1);
   try {
-    result = (PyObject *)((swig::SwigPyIterator const *)arg1)->value();
+    {
+      SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+      result = (PyObject *)((swig::SwigPyIterator const *)arg1)->value();
+      SWIG_PYTHON_THREAD_END_ALLOW;
+    }
   } catch(swig::stop_iteration &_e) {
     {
       (void)_e;
@@ -5004,7 +5013,11 @@ SWIGINTERN PyObject *_wrap_SwigPyIterator_incr__SWIG_0(PyObject *SWIGUNUSEDPARM(
   } 
   arg2 = static_cast< size_t >(val2);
   try {
-    result = (swig::SwigPyIterator *)(arg1)->incr(arg2);
+    {
+      SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+      result = (swig::SwigPyIterator *)(arg1)->incr(arg2);
+      SWIG_PYTHON_THREAD_END_ALLOW;
+    }
   } catch(swig::stop_iteration &_e) {
     {
       (void)_e;
@@ -5033,7 +5046,11 @@ SWIGINTERN PyObject *_wrap_SwigPyIterator_incr__SWIG_1(PyObject *SWIGUNUSEDPARM(
   }
   arg1 = reinterpret_cast< swig::SwigPyIterator * >(argp1);
   try {
-    result = (swig::SwigPyIterator *)(arg1)->incr();
+    {
+      SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+      result = (swig::SwigPyIterator *)(arg1)->incr();
+      SWIG_PYTHON_THREAD_END_ALLOW;
+    }
   } catch(swig::stop_iteration &_e) {
     {
       (void)_e;
@@ -5112,7 +5129,11 @@ SWIGINTERN PyObject *_wrap_SwigPyIterator_decr__SWIG_0(PyObject *SWIGUNUSEDPARM(
   } 
   arg2 = static_cast< size_t >(val2);
   try {
-    result = (swig::SwigPyIterator *)(arg1)->decr(arg2);
+    {
+      SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+      result = (swig::SwigPyIterator *)(arg1)->decr(arg2);
+      SWIG_PYTHON_THREAD_END_ALLOW;
+    }
   } catch(swig::stop_iteration &_e) {
     {
       (void)_e;
@@ -5141,7 +5162,11 @@ SWIGINTERN PyObject *_wrap_SwigPyIterator_decr__SWIG_1(PyObject *SWIGUNUSEDPARM(
   }
   arg1 = reinterpret_cast< swig::SwigPyIterator * >(argp1);
   try {
-    result = (swig::SwigPyIterator *)(arg1)->decr();
+    {
+      SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+      result = (swig::SwigPyIterator *)(arg1)->decr();
+      SWIG_PYTHON_THREAD_END_ALLOW;
+    }
   } catch(swig::stop_iteration &_e) {
     {
       (void)_e;
@@ -5224,7 +5249,11 @@ SWIGINTERN PyObject *_wrap_SwigPyIterator_distance(PyObject *SWIGUNUSEDPARM(self
   }
   arg2 = reinterpret_cast< swig::SwigPyIterator * >(argp2);
   try {
-    result = ((swig::SwigPyIterator const *)arg1)->distance((swig::SwigPyIterator const &)*arg2);
+    {
+      SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+      result = ((swig::SwigPyIterator const *)arg1)->distance((swig::SwigPyIterator const &)*arg2);
+      SWIG_PYTHON_THREAD_END_ALLOW;
+    }
   } catch(std::invalid_argument &_e) {
     SWIG_Python_Raise(SWIG_NewPointerObj((new std::invalid_argument(static_cast< const std::invalid_argument& >(_e))),SWIGTYPE_p_std__invalid_argument,SWIG_POINTER_OWN), "std::invalid_argument", SWIGTYPE_p_std__invalid_argument); SWIG_fail;
   }
@@ -5261,7 +5290,11 @@ SWIGINTERN PyObject *_wrap_SwigPyIterator_equal(PyObject *SWIGUNUSEDPARM(self), 
   }
   arg2 = reinterpret_cast< swig::SwigPyIterator * >(argp2);
   try {
-    result = (bool)((swig::SwigPyIterator const *)arg1)->equal((swig::SwigPyIterator const &)*arg2);
+    {
+      SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+      result = (bool)((swig::SwigPyIterator const *)arg1)->equal((swig::SwigPyIterator const &)*arg2);
+      SWIG_PYTHON_THREAD_END_ALLOW;
+    }
   } catch(std::invalid_argument &_e) {
     SWIG_Python_Raise(SWIG_NewPointerObj((new std::invalid_argument(static_cast< const std::invalid_argument& >(_e))),SWIGTYPE_p_std__invalid_argument,SWIG_POINTER_OWN), "std::invalid_argument", SWIGTYPE_p_std__invalid_argument); SWIG_fail;
   }
@@ -5287,7 +5320,11 @@ SWIGINTERN PyObject *_wrap_SwigPyIterator_copy(PyObject *SWIGUNUSEDPARM(self), P
     SWIG_exception_fail(SWIG_ArgError(res1), "in method '" "SwigPyIterator_copy" "', argument " "1"" of type '" "swig::SwigPyIterator const *""'"); 
   }
   arg1 = reinterpret_cast< swig::SwigPyIterator * >(argp1);
-  result = (swig::SwigPyIterator *)((swig::SwigPyIterator const *)arg1)->copy();
+  {
+    SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+    result = (swig::SwigPyIterator *)((swig::SwigPyIterator const *)arg1)->copy();
+    SWIG_PYTHON_THREAD_END_ALLOW;
+  }
   resultobj = SWIG_NewPointerObj(SWIG_as_voidptr(result), SWIGTYPE_p_swig__SwigPyIterator, SWIG_POINTER_OWN |  0 );
   return resultobj;
 fail:
@@ -5311,7 +5348,11 @@ SWIGINTERN PyObject *_wrap_SwigPyIterator_next(PyObject *SWIGUNUSEDPARM(self), P
   }
   arg1 = reinterpret_cast< swig::SwigPyIterator * >(argp1);
   try {
-    result = (PyObject *)(arg1)->next();
+    {
+      SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+      result = (PyObject *)(arg1)->next();
+      SWIG_PYTHON_THREAD_END_ALLOW;
+    }
   } catch(swig::stop_iteration &_e) {
     {
       (void)_e;
@@ -5342,7 +5383,11 @@ SWIGINTERN PyObject *_wrap_SwigPyIterator___next__(PyObject *SWIGUNUSEDPARM(self
   }
   arg1 = reinterpret_cast< swig::SwigPyIterator * >(argp1);
   try {
-    result = (PyObject *)(arg1)->__next__();
+    {
+      SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+      result = (PyObject *)(arg1)->__next__();
+      SWIG_PYTHON_THREAD_END_ALLOW;
+    }
   } catch(swig::stop_iteration &_e) {
     {
       (void)_e;
@@ -5373,7 +5418,11 @@ SWIGINTERN PyObject *_wrap_SwigPyIterator_previous(PyObject *SWIGUNUSEDPARM(self
   }
   arg1 = reinterpret_cast< swig::SwigPyIterator * >(argp1);
   try {
-    result = (PyObject *)(arg1)->previous();
+    {
+      SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+      result = (PyObject *)(arg1)->previous();
+      SWIG_PYTHON_THREAD_END_ALLOW;
+    }
   } catch(swig::stop_iteration &_e) {
     {
       (void)_e;
@@ -5411,7 +5460,11 @@ SWIGINTERN PyObject *_wrap_SwigPyIterator_advance(PyObject *SWIGUNUSEDPARM(self)
   } 
   arg2 = static_cast< ptrdiff_t >(val2);
   try {
-    result = (swig::SwigPyIterator *)(arg1)->advance(arg2);
+    {
+      SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+      result = (swig::SwigPyIterator *)(arg1)->advance(arg2);
+      SWIG_PYTHON_THREAD_END_ALLOW;
+    }
   } catch(swig::stop_iteration &_e) {
     {
       (void)_e;
@@ -5451,7 +5504,11 @@ SWIGINTERN PyObject *_wrap_SwigPyIterator___eq__(PyObject *SWIGUNUSEDPARM(self),
     SWIG_exception_fail(SWIG_ValueError, "invalid null reference " "in method '" "SwigPyIterator___eq__" "', argument " "2"" of type '" "swig::SwigPyIterator const &""'"); 
   }
   arg2 = reinterpret_cast< swig::SwigPyIterator * >(argp2);
-  result = (bool)((swig::SwigPyIterator const *)arg1)->operator ==((swig::SwigPyIterator const &)*arg2);
+  {
+    SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+    result = (bool)((swig::SwigPyIterator const *)arg1)->operator ==((swig::SwigPyIterator const &)*arg2);
+    SWIG_PYTHON_THREAD_END_ALLOW;
+  }
   resultobj = SWIG_From_bool(static_cast< bool >(result));
   return resultobj;
 fail:
@@ -5486,7 +5543,11 @@ SWIGINTERN PyObject *_wrap_SwigPyIterator___ne__(PyObject *SWIGUNUSEDPARM(self),
     SWIG_exception_fail(SWIG_ValueError, "invalid null reference " "in method '" "SwigPyIterator___ne__" "', argument " "2"" of type '" "swig::SwigPyIterator const &""'"); 
   }
   arg2 = reinterpret_cast< swig::SwigPyIterator * >(argp2);
-  result = (bool)((swig::SwigPyIterator const *)arg1)->operator !=((swig::SwigPyIterator const &)*arg2);
+  {
+    SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+    result = (bool)((swig::SwigPyIterator const *)arg1)->operator !=((swig::SwigPyIterator const &)*arg2);
+    SWIG_PYTHON_THREAD_END_ALLOW;
+  }
   resultobj = SWIG_From_bool(static_cast< bool >(result));
   return resultobj;
 fail:
@@ -5519,7 +5580,11 @@ SWIGINTERN PyObject *_wrap_SwigPyIterator___iadd__(PyObject *SWIGUNUSEDPARM(self
   } 
   arg2 = static_cast< ptrdiff_t >(val2);
   try {
-    result = (swig::SwigPyIterator *) &(arg1)->operator +=(arg2);
+    {
+      SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+      result = (swig::SwigPyIterator *) &(arg1)->operator +=(arg2);
+      SWIG_PYTHON_THREAD_END_ALLOW;
+    }
   } catch(swig::stop_iteration &_e) {
     {
       (void)_e;
@@ -5557,7 +5622,11 @@ SWIGINTERN PyObject *_wrap_SwigPyIterator___isub__(PyObject *SWIGUNUSEDPARM(self
   } 
   arg2 = static_cast< ptrdiff_t >(val2);
   try {
-    result = (swig::SwigPyIterator *) &(arg1)->operator -=(arg2);
+    {
+      SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+      result = (swig::SwigPyIterator *) &(arg1)->operator -=(arg2);
+      SWIG_PYTHON_THREAD_END_ALLOW;
+    }
   } catch(swig::stop_iteration &_e) {
     {
       (void)_e;
@@ -5595,7 +5664,11 @@ SWIGINTERN PyObject *_wrap_SwigPyIterator___add__(PyObject *SWIGUNUSEDPARM(self)
   } 
   arg2 = static_cast< ptrdiff_t >(val2);
   try {
-    result = (swig::SwigPyIterator *)((swig::SwigPyIterator const *)arg1)->operator +(arg2);
+    {
+      SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+      result = (swig::SwigPyIterator *)((swig::SwigPyIterator const *)arg1)->operator +(arg2);
+      SWIG_PYTHON_THREAD_END_ALLOW;
+    }
   } catch(swig::stop_iteration &_e) {
     {
       (void)_e;
@@ -5634,7 +5707,11 @@ SWIGINTERN PyObject *_wrap_SwigPyIterator___sub____SWIG_0(PyObject *SWIGUNUSEDPA
   } 
   arg2 = static_cast< ptrdiff_t >(val2);
   try {
-    result = (swig::SwigPyIterator *)((swig::SwigPyIterator const *)arg1)->operator -(arg2);
+    {
+      SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+      result = (swig::SwigPyIterator *)((swig::SwigPyIterator const *)arg1)->operator -(arg2);
+      SWIG_PYTHON_THREAD_END_ALLOW;
+    }
   } catch(swig::stop_iteration &_e) {
     {
       (void)_e;
@@ -5675,7 +5752,11 @@ SWIGINTERN PyObject *_wrap_SwigPyIterator___sub____SWIG_1(PyObject *SWIGUNUSEDPA
     SWIG_exception_fail(SWIG_ValueError, "invalid null reference " "in method '" "SwigPyIterator___sub__" "', argument " "2"" of type '" "swig::SwigPyIterator const &""'"); 
   }
   arg2 = reinterpret_cast< swig::SwigPyIterator * >(argp2);
-  result = ((swig::SwigPyIterator const *)arg1)->operator -((swig::SwigPyIterator const &)*arg2);
+  {
+    SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+    result = ((swig::SwigPyIterator const *)arg1)->operator -((swig::SwigPyIterator const &)*arg2);
+    SWIG_PYTHON_THREAD_END_ALLOW;
+  }
   resultobj = SWIG_From_ptrdiff_t(static_cast< ptrdiff_t >(result));
   return resultobj;
 fail:
@@ -5751,7 +5832,11 @@ SWIGINTERN PyObject *_wrap_delete_BTSerialPortBinding(PyObject *SWIGUNUSEDPARM(s
   arg1 = reinterpret_cast< BTSerialPortBinding * >(argp1);
   {
     try {
-      delete arg1;
+      {
+        SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+        delete arg1;
+        SWIG_PYTHON_THREAD_END_ALLOW;
+      }
     } catch(const ExploreException& e) {
       SWIG_exception(SWIG_ValueError, e.what());
     }
@@ -5797,7 +5882,11 @@ SWIGINTERN PyObject *_wrap_BTSerialPortBinding_Create(PyObject *SWIGUNUSEDPARM(s
   arg2 = static_cast< int >(val2);
   {
     try {
-      result = (BTSerialPortBinding *)BTSerialPortBinding::Create(arg1,arg2);
+      {
+        SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+        result = (BTSerialPortBinding *)BTSerialPortBinding::Create(arg1,arg2);
+        SWIG_PYTHON_THREAD_END_ALLOW;
+      }
     } catch(const ExploreException& e) {
       SWIG_exception(SWIG_ValueError, e.what());
     }
@@ -5834,7 +5923,11 @@ SWIGINTERN PyObject *_wrap_BTSerialPortBinding_Connect(PyObject *SWIGUNUSEDPARM(
   arg1 = reinterpret_cast< BTSerialPortBinding * >(argp1);
   {
     try {
-      result = (int)(arg1)->Connect();
+      {
+        SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+        result = (int)(arg1)->Connect();
+        SWIG_PYTHON_THREAD_END_ALLOW;
+      }
     } catch(const ExploreException& e) {
       SWIG_exception(SWIG_ValueError, e.what());
     }
@@ -5870,7 +5963,11 @@ SWIGINTERN PyObject *_wrap_BTSerialPortBinding_Close(PyObject *SWIGUNUSEDPARM(se
   arg1 = reinterpret_cast< BTSerialPortBinding * >(argp1);
   {
     try {
-      (arg1)->Close();
+      {
+        SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+        (arg1)->Close();
+        SWIG_PYTHON_THREAD_END_ALLOW;
+      }
     } catch(const ExploreException& e) {
       SWIG_exception(SWIG_ValueError, e.what());
     }
@@ -5919,7 +6016,11 @@ SWIGINTERN PyObject *_wrap_BTSerialPortBinding_Read(PyObject *SWIGUNUSEDPARM(sel
   arg3 = &size2;
   {
     try {
-      (arg1)->Read(arg2,arg3);
+      {
+        SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+        (arg1)->Read(arg2,arg3);
+        SWIG_PYTHON_THREAD_END_ALLOW;
+      }
     } catch(const ExploreException& e) {
       SWIG_exception(SWIG_ValueError, e.what());
     }
@@ -5964,7 +6065,11 @@ SWIGINTERN PyObject *_wrap_BTSerialPortBinding_Write(PyObject *SWIGUNUSEDPARM(se
   }
   {
     try {
-      (arg1)->Write((char const *)arg2,arg3);
+      {
+        SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+        (arg1)->Write((char const *)arg2,arg3);
+        SWIG_PYTHON_THREAD_END_ALLOW;
+      }
     } catch(const ExploreException& e) {
       SWIG_exception(SWIG_ValueError, e.what());
     }
@@ -6001,7 +6106,11 @@ SWIGINTERN PyObject *_wrap_BTSerialPortBinding_IsDataAvailable(PyObject *SWIGUNU
   arg1 = reinterpret_cast< BTSerialPortBinding * >(argp1);
   {
     try {
-      result = (bool)(arg1)->IsDataAvailable();
+      {
+        SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+        result = (bool)(arg1)->IsDataAvailable();
+        SWIG_PYTHON_THREAD_END_ALLOW;
+      }
     } catch(const ExploreException& e) {
       SWIG_exception(SWIG_ValueError, e.what());
     }
@@ -6047,7 +6156,11 @@ SWIGINTERN PyObject *_wrap_vectordevice_iterator(PyObject *SWIGUNUSEDPARM(self),
   arg1 = reinterpret_cast< std::vector< device > * >(argp1);
   {
     try {
-      result = (swig::SwigPyIterator *)std_vector_Sl_device_Sg__iterator(arg1,arg2);
+      {
+        SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+        result = (swig::SwigPyIterator *)std_vector_Sl_device_Sg__iterator(arg1,arg2);
+        SWIG_PYTHON_THREAD_END_ALLOW;
+      }
     } catch(const ExploreException& e) {
       SWIG_exception(SWIG_ValueError, e.what());
     }
@@ -6084,7 +6197,11 @@ SWIGINTERN PyObject *_wrap_vectordevice___nonzero__(PyObject *SWIGUNUSEDPARM(sel
   arg1 = reinterpret_cast< std::vector< device > * >(argp1);
   {
     try {
-      result = (bool)std_vector_Sl_device_Sg____nonzero__((std::vector< device > const *)arg1);
+      {
+        SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+        result = (bool)std_vector_Sl_device_Sg____nonzero__((std::vector< device > const *)arg1);
+        SWIG_PYTHON_THREAD_END_ALLOW;
+      }
     } catch(const ExploreException& e) {
       SWIG_exception(SWIG_ValueError, e.what());
     }
@@ -6121,7 +6238,11 @@ SWIGINTERN PyObject *_wrap_vectordevice___bool__(PyObject *SWIGUNUSEDPARM(self),
   arg1 = reinterpret_cast< std::vector< device > * >(argp1);
   {
     try {
-      result = (bool)std_vector_Sl_device_Sg____bool__((std::vector< device > const *)arg1);
+      {
+        SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+        result = (bool)std_vector_Sl_device_Sg____bool__((std::vector< device > const *)arg1);
+        SWIG_PYTHON_THREAD_END_ALLOW;
+      }
     } catch(const ExploreException& e) {
       SWIG_exception(SWIG_ValueError, e.what());
     }
@@ -6158,7 +6279,11 @@ SWIGINTERN PyObject *_wrap_vectordevice___len__(PyObject *SWIGUNUSEDPARM(self), 
   arg1 = reinterpret_cast< std::vector< device > * >(argp1);
   {
     try {
-      result = std_vector_Sl_device_Sg____len__((std::vector< device > const *)arg1);
+      {
+        SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+        result = std_vector_Sl_device_Sg____len__((std::vector< device > const *)arg1);
+        SWIG_PYTHON_THREAD_END_ALLOW;
+      }
     } catch(const ExploreException& e) {
       SWIG_exception(SWIG_ValueError, e.what());
     }
@@ -6211,7 +6336,11 @@ SWIGINTERN PyObject *_wrap_vectordevice___getslice__(PyObject *SWIGUNUSEDPARM(se
   {
     try {
       try {
-        result = (std::vector< device,std::allocator< device > > *)std_vector_Sl_device_Sg____getslice__(arg1,arg2,arg3);
+        {
+          SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+          result = (std::vector< device,std::allocator< device > > *)std_vector_Sl_device_Sg____getslice__(arg1,arg2,arg3);
+          SWIG_PYTHON_THREAD_END_ALLOW;
+        }
       } catch(std::out_of_range &_e) {
         SWIG_exception_fail(SWIG_IndexError, (&_e)->what());
       } catch(std::invalid_argument &_e) {
@@ -6267,7 +6396,11 @@ SWIGINTERN PyObject *_wrap_vectordevice___setslice____SWIG_0(PyObject *SWIGUNUSE
   {
     try {
       try {
-        std_vector_Sl_device_Sg____setslice____SWIG_0(arg1,arg2,arg3);
+        {
+          SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+          std_vector_Sl_device_Sg____setslice____SWIG_0(arg1,arg2,arg3);
+          SWIG_PYTHON_THREAD_END_ALLOW;
+        }
       } catch(std::out_of_range &_e) {
         SWIG_exception_fail(SWIG_IndexError, (&_e)->what());
       } catch(std::invalid_argument &_e) {
@@ -6336,7 +6469,11 @@ SWIGINTERN PyObject *_wrap_vectordevice___setslice____SWIG_1(PyObject *SWIGUNUSE
   {
     try {
       try {
-        std_vector_Sl_device_Sg____setslice____SWIG_1(arg1,arg2,arg3,(std::vector< device,std::allocator< device > > const &)*arg4);
+        {
+          SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+          std_vector_Sl_device_Sg____setslice____SWIG_1(arg1,arg2,arg3,(std::vector< device,std::allocator< device > > const &)*arg4);
+          SWIG_PYTHON_THREAD_END_ALLOW;
+        }
       } catch(std::out_of_range &_e) {
         SWIG_exception_fail(SWIG_IndexError, (&_e)->what());
       } catch(std::invalid_argument &_e) {
@@ -6457,7 +6594,11 @@ SWIGINTERN PyObject *_wrap_vectordevice___delslice__(PyObject *SWIGUNUSEDPARM(se
   {
     try {
       try {
-        std_vector_Sl_device_Sg____delslice__(arg1,arg2,arg3);
+        {
+          SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+          std_vector_Sl_device_Sg____delslice__(arg1,arg2,arg3);
+          SWIG_PYTHON_THREAD_END_ALLOW;
+        }
       } catch(std::out_of_range &_e) {
         SWIG_exception_fail(SWIG_IndexError, (&_e)->what());
       } catch(std::invalid_argument &_e) {
@@ -6505,7 +6646,11 @@ SWIGINTERN PyObject *_wrap_vectordevice___delitem____SWIG_0(PyObject *SWIGUNUSED
   {
     try {
       try {
-        std_vector_Sl_device_Sg____delitem____SWIG_0(arg1,arg2);
+        {
+          SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+          std_vector_Sl_device_Sg____delitem____SWIG_0(arg1,arg2);
+          SWIG_PYTHON_THREAD_END_ALLOW;
+        }
       } catch(std::out_of_range &_e) {
         SWIG_exception_fail(SWIG_IndexError, (&_e)->what());
       } catch(std::invalid_argument &_e) {
@@ -6553,7 +6698,11 @@ SWIGINTERN PyObject *_wrap_vectordevice___getitem____SWIG_0(PyObject *SWIGUNUSED
   {
     try {
       try {
-        result = (std::vector< device,std::allocator< device > > *)std_vector_Sl_device_Sg____getitem____SWIG_0(arg1,arg2);
+        {
+          SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+          result = (std::vector< device,std::allocator< device > > *)std_vector_Sl_device_Sg____getitem____SWIG_0(arg1,arg2);
+          SWIG_PYTHON_THREAD_END_ALLOW;
+        }
       } catch(std::out_of_range &_e) {
         SWIG_exception_fail(SWIG_IndexError, (&_e)->what());
       } catch(std::invalid_argument &_e) {
@@ -6613,7 +6762,11 @@ SWIGINTERN PyObject *_wrap_vectordevice___setitem____SWIG_0(PyObject *SWIGUNUSED
   {
     try {
       try {
-        std_vector_Sl_device_Sg____setitem____SWIG_0(arg1,arg2,(std::vector< device,std::allocator< device > > const &)*arg3);
+        {
+          SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+          std_vector_Sl_device_Sg____setitem____SWIG_0(arg1,arg2,(std::vector< device,std::allocator< device > > const &)*arg3);
+          SWIG_PYTHON_THREAD_END_ALLOW;
+        }
       } catch(std::out_of_range &_e) {
         SWIG_exception_fail(SWIG_IndexError, (&_e)->what());
       } catch(std::invalid_argument &_e) {
@@ -6662,7 +6815,11 @@ SWIGINTERN PyObject *_wrap_vectordevice___setitem____SWIG_1(PyObject *SWIGUNUSED
   {
     try {
       try {
-        std_vector_Sl_device_Sg____setitem____SWIG_1(arg1,arg2);
+        {
+          SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+          std_vector_Sl_device_Sg____setitem____SWIG_1(arg1,arg2);
+          SWIG_PYTHON_THREAD_END_ALLOW;
+        }
       } catch(std::out_of_range &_e) {
         SWIG_exception_fail(SWIG_IndexError, (&_e)->what());
       } catch(std::invalid_argument &_e) {
@@ -6709,7 +6866,11 @@ SWIGINTERN PyObject *_wrap_vectordevice___delitem____SWIG_1(PyObject *SWIGUNUSED
   {
     try {
       try {
-        std_vector_Sl_device_Sg____delitem____SWIG_1(arg1,arg2);
+        {
+          SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+          std_vector_Sl_device_Sg____delitem____SWIG_1(arg1,arg2);
+          SWIG_PYTHON_THREAD_END_ALLOW;
+        }
       } catch(std::out_of_range &_e) {
         SWIG_exception_fail(SWIG_IndexError, (&_e)->what());
       } catch(std::invalid_argument &_e) {
@@ -6803,7 +6964,11 @@ SWIGINTERN PyObject *_wrap_vectordevice___getitem____SWIG_1(PyObject *SWIGUNUSED
   {
     try {
       try {
-        result = (std::vector< device >::value_type *) &std_vector_Sl_device_Sg____getitem____SWIG_1((std::vector< device > const *)arg1,arg2);
+        {
+          SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+          result = (std::vector< device >::value_type *) &std_vector_Sl_device_Sg____getitem____SWIG_1((std::vector< device > const *)arg1,arg2);
+          SWIG_PYTHON_THREAD_END_ALLOW;
+        }
       } catch(std::out_of_range &_e) {
         SWIG_exception_fail(SWIG_IndexError, (&_e)->what());
       }
@@ -6906,7 +7071,11 @@ SWIGINTERN PyObject *_wrap_vectordevice___setitem____SWIG_2(PyObject *SWIGUNUSED
   {
     try {
       try {
-        std_vector_Sl_device_Sg____setitem____SWIG_2(arg1,arg2,(device const &)*arg3);
+        {
+          SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+          std_vector_Sl_device_Sg____setitem____SWIG_2(arg1,arg2,(device const &)*arg3);
+          SWIG_PYTHON_THREAD_END_ALLOW;
+        }
       } catch(std::out_of_range &_e) {
         SWIG_exception_fail(SWIG_IndexError, (&_e)->what());
       }
@@ -7014,7 +7183,11 @@ SWIGINTERN PyObject *_wrap_vectordevice_pop(PyObject *SWIGUNUSEDPARM(self), PyOb
   {
     try {
       try {
-        result = std_vector_Sl_device_Sg__pop(arg1);
+        {
+          SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+          result = std_vector_Sl_device_Sg__pop(arg1);
+          SWIG_PYTHON_THREAD_END_ALLOW;
+        }
       } catch(std::out_of_range &_e) {
         SWIG_exception_fail(SWIG_IndexError, (&_e)->what());
       }
@@ -7063,7 +7236,11 @@ SWIGINTERN PyObject *_wrap_vectordevice_append(PyObject *SWIGUNUSEDPARM(self), P
   arg2 = reinterpret_cast< std::vector< device >::value_type * >(argp2);
   {
     try {
-      std_vector_Sl_device_Sg__append(arg1,(device const &)*arg2);
+      {
+        SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+        std_vector_Sl_device_Sg__append(arg1,(device const &)*arg2);
+        SWIG_PYTHON_THREAD_END_ALLOW;
+      }
     } catch(const ExploreException& e) {
       SWIG_exception(SWIG_ValueError, e.what());
     }
@@ -7090,7 +7267,11 @@ SWIGINTERN PyObject *_wrap_new_vectordevice__SWIG_0(PyObject *SWIGUNUSEDPARM(sel
   if ((nobjs < 0) || (nobjs > 0)) SWIG_fail;
   {
     try {
-      result = (std::vector< device > *)new std::vector< device >();
+      {
+        SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+        result = (std::vector< device > *)new std::vector< device >();
+        SWIG_PYTHON_THREAD_END_ALLOW;
+      }
     } catch(const ExploreException& e) {
       SWIG_exception(SWIG_ValueError, e.what());
     }
@@ -7130,7 +7311,11 @@ SWIGINTERN PyObject *_wrap_new_vectordevice__SWIG_1(PyObject *SWIGUNUSEDPARM(sel
   }
   {
     try {
-      result = (std::vector< device > *)new std::vector< device >((std::vector< device > const &)*arg1);
+      {
+        SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+        result = (std::vector< device > *)new std::vector< device >((std::vector< device > const &)*arg1);
+        SWIG_PYTHON_THREAD_END_ALLOW;
+      }
     } catch(const ExploreException& e) {
       SWIG_exception(SWIG_ValueError, e.what());
     }
@@ -7169,7 +7354,11 @@ SWIGINTERN PyObject *_wrap_vectordevice_empty(PyObject *SWIGUNUSEDPARM(self), Py
   arg1 = reinterpret_cast< std::vector< device > * >(argp1);
   {
     try {
-      result = (bool)((std::vector< device > const *)arg1)->empty();
+      {
+        SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+        result = (bool)((std::vector< device > const *)arg1)->empty();
+        SWIG_PYTHON_THREAD_END_ALLOW;
+      }
     } catch(const ExploreException& e) {
       SWIG_exception(SWIG_ValueError, e.what());
     }
@@ -7206,7 +7395,11 @@ SWIGINTERN PyObject *_wrap_vectordevice_size(PyObject *SWIGUNUSEDPARM(self), PyO
   arg1 = reinterpret_cast< std::vector< device > * >(argp1);
   {
     try {
-      result = ((std::vector< device > const *)arg1)->size();
+      {
+        SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+        result = ((std::vector< device > const *)arg1)->size();
+        SWIG_PYTHON_THREAD_END_ALLOW;
+      }
     } catch(const ExploreException& e) {
       SWIG_exception(SWIG_ValueError, e.what());
     }
@@ -7252,7 +7445,11 @@ SWIGINTERN PyObject *_wrap_vectordevice_swap(PyObject *SWIGUNUSEDPARM(self), PyO
   arg2 = reinterpret_cast< std::vector< device > * >(argp2);
   {
     try {
-      (arg1)->swap(*arg2);
+      {
+        SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+        (arg1)->swap(*arg2);
+        SWIG_PYTHON_THREAD_END_ALLOW;
+      }
     } catch(const ExploreException& e) {
       SWIG_exception(SWIG_ValueError, e.what());
     }
@@ -7289,7 +7486,11 @@ SWIGINTERN PyObject *_wrap_vectordevice_begin(PyObject *SWIGUNUSEDPARM(self), Py
   arg1 = reinterpret_cast< std::vector< device > * >(argp1);
   {
     try {
-      result = (arg1)->begin();
+      {
+        SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+        result = (arg1)->begin();
+        SWIG_PYTHON_THREAD_END_ALLOW;
+      }
     } catch(const ExploreException& e) {
       SWIG_exception(SWIG_ValueError, e.what());
     }
@@ -7327,7 +7528,11 @@ SWIGINTERN PyObject *_wrap_vectordevice_end(PyObject *SWIGUNUSEDPARM(self), PyOb
   arg1 = reinterpret_cast< std::vector< device > * >(argp1);
   {
     try {
-      result = (arg1)->end();
+      {
+        SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+        result = (arg1)->end();
+        SWIG_PYTHON_THREAD_END_ALLOW;
+      }
     } catch(const ExploreException& e) {
       SWIG_exception(SWIG_ValueError, e.what());
     }
@@ -7365,7 +7570,11 @@ SWIGINTERN PyObject *_wrap_vectordevice_rbegin(PyObject *SWIGUNUSEDPARM(self), P
   arg1 = reinterpret_cast< std::vector< device > * >(argp1);
   {
     try {
-      result = (arg1)->rbegin();
+      {
+        SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+        result = (arg1)->rbegin();
+        SWIG_PYTHON_THREAD_END_ALLOW;
+      }
     } catch(const ExploreException& e) {
       SWIG_exception(SWIG_ValueError, e.what());
     }
@@ -7403,7 +7612,11 @@ SWIGINTERN PyObject *_wrap_vectordevice_rend(PyObject *SWIGUNUSEDPARM(self), PyO
   arg1 = reinterpret_cast< std::vector< device > * >(argp1);
   {
     try {
-      result = (arg1)->rend();
+      {
+        SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+        result = (arg1)->rend();
+        SWIG_PYTHON_THREAD_END_ALLOW;
+      }
     } catch(const ExploreException& e) {
       SWIG_exception(SWIG_ValueError, e.what());
     }
@@ -7440,7 +7653,11 @@ SWIGINTERN PyObject *_wrap_vectordevice_clear(PyObject *SWIGUNUSEDPARM(self), Py
   arg1 = reinterpret_cast< std::vector< device > * >(argp1);
   {
     try {
-      (arg1)->clear();
+      {
+        SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+        (arg1)->clear();
+        SWIG_PYTHON_THREAD_END_ALLOW;
+      }
     } catch(const ExploreException& e) {
       SWIG_exception(SWIG_ValueError, e.what());
     }
@@ -7477,7 +7694,11 @@ SWIGINTERN PyObject *_wrap_vectordevice_get_allocator(PyObject *SWIGUNUSEDPARM(s
   arg1 = reinterpret_cast< std::vector< device > * >(argp1);
   {
     try {
-      result = ((std::vector< device > const *)arg1)->get_allocator();
+      {
+        SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+        result = ((std::vector< device > const *)arg1)->get_allocator();
+        SWIG_PYTHON_THREAD_END_ALLOW;
+      }
     } catch(const ExploreException& e) {
       SWIG_exception(SWIG_ValueError, e.what());
     }
@@ -7512,7 +7733,11 @@ SWIGINTERN PyObject *_wrap_new_vectordevice__SWIG_2(PyObject *SWIGUNUSEDPARM(sel
   arg1 = static_cast< std::vector< device >::size_type >(val1);
   {
     try {
-      result = (std::vector< device > *)new std::vector< device >(arg1);
+      {
+        SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+        result = (std::vector< device > *)new std::vector< device >(arg1);
+        SWIG_PYTHON_THREAD_END_ALLOW;
+      }
     } catch(const ExploreException& e) {
       SWIG_exception(SWIG_ValueError, e.what());
     }
@@ -7548,7 +7773,11 @@ SWIGINTERN PyObject *_wrap_vectordevice_pop_back(PyObject *SWIGUNUSEDPARM(self),
   arg1 = reinterpret_cast< std::vector< device > * >(argp1);
   {
     try {
-      (arg1)->pop_back();
+      {
+        SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+        (arg1)->pop_back();
+        SWIG_PYTHON_THREAD_END_ALLOW;
+      }
     } catch(const ExploreException& e) {
       SWIG_exception(SWIG_ValueError, e.what());
     }
@@ -7590,7 +7819,11 @@ SWIGINTERN PyObject *_wrap_vectordevice_resize__SWIG_0(PyObject *SWIGUNUSEDPARM(
   arg2 = static_cast< std::vector< device >::size_type >(val2);
   {
     try {
-      (arg1)->resize(arg2);
+      {
+        SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+        (arg1)->resize(arg2);
+        SWIG_PYTHON_THREAD_END_ALLOW;
+      }
     } catch(const ExploreException& e) {
       SWIG_exception(SWIG_ValueError, e.what());
     }
@@ -7639,7 +7872,11 @@ SWIGINTERN PyObject *_wrap_vectordevice_erase__SWIG_0(PyObject *SWIGUNUSEDPARM(s
   }
   {
     try {
-      result = std_vector_Sl_device_Sg__erase__SWIG_0(arg1,arg2);
+      {
+        SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+        result = std_vector_Sl_device_Sg__erase__SWIG_0(arg1,arg2);
+        SWIG_PYTHON_THREAD_END_ALLOW;
+      }
     } catch(const ExploreException& e) {
       SWIG_exception(SWIG_ValueError, e.what());
     }
@@ -7703,7 +7940,11 @@ SWIGINTERN PyObject *_wrap_vectordevice_erase__SWIG_1(PyObject *SWIGUNUSEDPARM(s
   }
   {
     try {
-      result = std_vector_Sl_device_Sg__erase__SWIG_1(arg1,arg2,arg3);
+      {
+        SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+        result = std_vector_Sl_device_Sg__erase__SWIG_1(arg1,arg2,arg3);
+        SWIG_PYTHON_THREAD_END_ALLOW;
+      }
     } catch(const ExploreException& e) {
       SWIG_exception(SWIG_ValueError, e.what());
     }
@@ -7799,7 +8040,11 @@ SWIGINTERN PyObject *_wrap_new_vectordevice__SWIG_3(PyObject *SWIGUNUSEDPARM(sel
   arg2 = reinterpret_cast< std::vector< device >::value_type * >(argp2);
   {
     try {
-      result = (std::vector< device > *)new std::vector< device >(arg1,(std::vector< device >::value_type const &)*arg2);
+      {
+        SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+        result = (std::vector< device > *)new std::vector< device >(arg1,(std::vector< device >::value_type const &)*arg2);
+        SWIG_PYTHON_THREAD_END_ALLOW;
+      }
     } catch(const ExploreException& e) {
       SWIG_exception(SWIG_ValueError, e.what());
     }
@@ -7900,7 +8145,11 @@ SWIGINTERN PyObject *_wrap_vectordevice_push_back(PyObject *SWIGUNUSEDPARM(self)
   arg2 = reinterpret_cast< std::vector< device >::value_type * >(argp2);
   {
     try {
-      (arg1)->push_back((std::vector< device >::value_type const &)*arg2);
+      {
+        SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+        (arg1)->push_back((std::vector< device >::value_type const &)*arg2);
+        SWIG_PYTHON_THREAD_END_ALLOW;
+      }
     } catch(const ExploreException& e) {
       SWIG_exception(SWIG_ValueError, e.what());
     }
@@ -7937,7 +8186,11 @@ SWIGINTERN PyObject *_wrap_vectordevice_front(PyObject *SWIGUNUSEDPARM(self), Py
   arg1 = reinterpret_cast< std::vector< device > * >(argp1);
   {
     try {
-      result = (std::vector< device >::value_type *) &((std::vector< device > const *)arg1)->front();
+      {
+        SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+        result = (std::vector< device >::value_type *) &((std::vector< device > const *)arg1)->front();
+        SWIG_PYTHON_THREAD_END_ALLOW;
+      }
     } catch(const ExploreException& e) {
       SWIG_exception(SWIG_ValueError, e.what());
     }
@@ -7975,7 +8228,11 @@ SWIGINTERN PyObject *_wrap_vectordevice_back(PyObject *SWIGUNUSEDPARM(self), PyO
   arg1 = reinterpret_cast< std::vector< device > * >(argp1);
   {
     try {
-      result = (std::vector< device >::value_type *) &((std::vector< device > const *)arg1)->back();
+      {
+        SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+        result = (std::vector< device >::value_type *) &((std::vector< device > const *)arg1)->back();
+        SWIG_PYTHON_THREAD_END_ALLOW;
+      }
     } catch(const ExploreException& e) {
       SWIG_exception(SWIG_ValueError, e.what());
     }
@@ -8030,7 +8287,11 @@ SWIGINTERN PyObject *_wrap_vectordevice_assign(PyObject *SWIGUNUSEDPARM(self), P
   arg3 = reinterpret_cast< std::vector< device >::value_type * >(argp3);
   {
     try {
-      (arg1)->assign(arg2,(std::vector< device >::value_type const &)*arg3);
+      {
+        SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+        (arg1)->assign(arg2,(std::vector< device >::value_type const &)*arg3);
+        SWIG_PYTHON_THREAD_END_ALLOW;
+      }
     } catch(const ExploreException& e) {
       SWIG_exception(SWIG_ValueError, e.what());
     }
@@ -8083,7 +8344,11 @@ SWIGINTERN PyObject *_wrap_vectordevice_resize__SWIG_1(PyObject *SWIGUNUSEDPARM(
   arg3 = reinterpret_cast< std::vector< device >::value_type * >(argp3);
   {
     try {
-      (arg1)->resize(arg2,(std::vector< device >::value_type const &)*arg3);
+      {
+        SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+        (arg1)->resize(arg2,(std::vector< device >::value_type const &)*arg3);
+        SWIG_PYTHON_THREAD_END_ALLOW;
+      }
     } catch(const ExploreException& e) {
       SWIG_exception(SWIG_ValueError, e.what());
     }
@@ -8193,7 +8458,11 @@ SWIGINTERN PyObject *_wrap_vectordevice_insert__SWIG_0(PyObject *SWIGUNUSEDPARM(
   arg3 = reinterpret_cast< std::vector< device >::value_type * >(argp3);
   {
     try {
-      result = std_vector_Sl_device_Sg__insert__SWIG_0(arg1,arg2,(device const &)*arg3);
+      {
+        SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+        result = std_vector_Sl_device_Sg__insert__SWIG_0(arg1,arg2,(device const &)*arg3);
+        SWIG_PYTHON_THREAD_END_ALLOW;
+      }
     } catch(const ExploreException& e) {
       SWIG_exception(SWIG_ValueError, e.what());
     }
@@ -8261,7 +8530,11 @@ SWIGINTERN PyObject *_wrap_vectordevice_insert__SWIG_1(PyObject *SWIGUNUSEDPARM(
   arg4 = reinterpret_cast< std::vector< device >::value_type * >(argp4);
   {
     try {
-      std_vector_Sl_device_Sg__insert__SWIG_1(arg1,arg2,arg3,(device const &)*arg4);
+      {
+        SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+        std_vector_Sl_device_Sg__insert__SWIG_1(arg1,arg2,arg3,(device const &)*arg4);
+        SWIG_PYTHON_THREAD_END_ALLOW;
+      }
     } catch(const ExploreException& e) {
       SWIG_exception(SWIG_ValueError, e.what());
     }
@@ -8362,7 +8635,11 @@ SWIGINTERN PyObject *_wrap_vectordevice_reserve(PyObject *SWIGUNUSEDPARM(self), 
   arg2 = static_cast< std::vector< device >::size_type >(val2);
   {
     try {
-      (arg1)->reserve(arg2);
+      {
+        SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+        (arg1)->reserve(arg2);
+        SWIG_PYTHON_THREAD_END_ALLOW;
+      }
     } catch(const ExploreException& e) {
       SWIG_exception(SWIG_ValueError, e.what());
     }
@@ -8399,7 +8676,11 @@ SWIGINTERN PyObject *_wrap_vectordevice_capacity(PyObject *SWIGUNUSEDPARM(self),
   arg1 = reinterpret_cast< std::vector< device > * >(argp1);
   {
     try {
-      result = ((std::vector< device > const *)arg1)->capacity();
+      {
+        SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+        result = ((std::vector< device > const *)arg1)->capacity();
+        SWIG_PYTHON_THREAD_END_ALLOW;
+      }
     } catch(const ExploreException& e) {
       SWIG_exception(SWIG_ValueError, e.what());
     }
@@ -8435,7 +8716,11 @@ SWIGINTERN PyObject *_wrap_delete_vectordevice(PyObject *SWIGUNUSEDPARM(self), P
   arg1 = reinterpret_cast< std::vector< device > * >(argp1);
   {
     try {
-      delete arg1;
+      {
+        SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+        delete arg1;
+        SWIG_PYTHON_THREAD_END_ALLOW;
+      }
     } catch(const ExploreException& e) {
       SWIG_exception(SWIG_ValueError, e.what());
     }
@@ -8485,7 +8770,11 @@ SWIGINTERN PyObject *_wrap_new_ExploreException(PyObject *SWIGUNUSEDPARM(self), 
   }
   {
     try {
-      result = (ExploreException *)new ExploreException(arg1);
+      {
+        SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+        result = (ExploreException *)new ExploreException(arg1);
+        SWIG_PYTHON_THREAD_END_ALLOW;
+      }
     } catch(const ExploreException& e) {
       SWIG_exception(SWIG_ValueError, e.what());
     }
@@ -8522,7 +8811,11 @@ SWIGINTERN PyObject *_wrap_ExploreException_what(PyObject *SWIGUNUSEDPARM(self),
   arg1 = reinterpret_cast< ExploreException * >(argp1);
   {
     try {
-      result = (char *)((ExploreException const *)arg1)->what();
+      {
+        SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+        result = (char *)((ExploreException const *)arg1)->what();
+        SWIG_PYTHON_THREAD_END_ALLOW;
+      }
     } catch(const ExploreException& e) {
       SWIG_exception(SWIG_ValueError, e.what());
     }
@@ -8558,7 +8851,11 @@ SWIGINTERN PyObject *_wrap_delete_ExploreException(PyObject *SWIGUNUSEDPARM(self
   arg1 = reinterpret_cast< ExploreException * >(argp1);
   {
     try {
-      delete arg1;
+      {
+        SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+        delete arg1;
+        SWIG_PYTHON_THREAD_END_ALLOW;
+      }
     } catch(const ExploreException& e) {
       SWIG_exception(SWIG_ValueError, e.what());
     }
@@ -8608,7 +8905,11 @@ SWIGINTERN PyObject *_wrap_new_ExploreReadBufferException(PyObject *SWIGUNUSEDPA
   }
   {
     try {
-      result = (ExploreReadBufferException *)new ExploreReadBufferException(arg1);
+      {
+        SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+        result = (ExploreReadBufferException *)new ExploreReadBufferException(arg1);
+        SWIG_PYTHON_THREAD_END_ALLOW;
+      }
     } catch(const ExploreException& e) {
       SWIG_exception(SWIG_ValueError, e.what());
     }
@@ -8645,7 +8946,11 @@ SWIGINTERN PyObject *_wrap_ExploreReadBufferException_what(PyObject *SWIGUNUSEDP
   arg1 = reinterpret_cast< ExploreReadBufferException * >(argp1);
   {
     try {
-      result = (char *)((ExploreReadBufferException const *)arg1)->what();
+      {
+        SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+        result = (char *)((ExploreReadBufferException const *)arg1)->what();
+        SWIG_PYTHON_THREAD_END_ALLOW;
+      }
     } catch(const ExploreException& e) {
       SWIG_exception(SWIG_ValueError, e.what());
     }
@@ -8681,7 +8986,11 @@ SWIGINTERN PyObject *_wrap_delete_ExploreReadBufferException(PyObject *SWIGUNUSE
   arg1 = reinterpret_cast< ExploreReadBufferException * >(argp1);
   {
     try {
-      delete arg1;
+      {
+        SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+        delete arg1;
+        SWIG_PYTHON_THREAD_END_ALLOW;
+      }
     } catch(const ExploreException& e) {
       SWIG_exception(SWIG_ValueError, e.what());
     }
@@ -8738,7 +9047,11 @@ SWIGINTERN PyObject *_wrap_device_address_set(PyObject *SWIGUNUSEDPARM(self), Py
     }
     arg2 = ptr;
   }
-  if (arg1) (arg1)->address = *arg2;
+  {
+    SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+    if (arg1) (arg1)->address = *arg2;
+    SWIG_PYTHON_THREAD_END_ALLOW;
+  }
   resultobj = SWIG_Py_Void();
   if (SWIG_IsNewObj(res2)) delete arg2;
   return resultobj;
@@ -8763,7 +9076,11 @@ SWIGINTERN PyObject *_wrap_device_address_get(PyObject *SWIGUNUSEDPARM(self), Py
     SWIG_exception_fail(SWIG_ArgError(res1), "in method '" "device_address_get" "', argument " "1"" of type '" "device *""'"); 
   }
   arg1 = reinterpret_cast< device * >(argp1);
-  result = (std::string *) & ((arg1)->address);
+  {
+    SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+    result = (std::string *) & ((arg1)->address);
+    SWIG_PYTHON_THREAD_END_ALLOW;
+  }
   resultobj = SWIG_From_std_string(static_cast< std::string >(*result));
   return resultobj;
 fail:
@@ -8797,7 +9114,11 @@ SWIGINTERN PyObject *_wrap_device_name_set(PyObject *SWIGUNUSEDPARM(self), PyObj
     }
     arg2 = ptr;
   }
-  if (arg1) (arg1)->name = *arg2;
+  {
+    SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+    if (arg1) (arg1)->name = *arg2;
+    SWIG_PYTHON_THREAD_END_ALLOW;
+  }
   resultobj = SWIG_Py_Void();
   if (SWIG_IsNewObj(res2)) delete arg2;
   return resultobj;
@@ -8822,7 +9143,11 @@ SWIGINTERN PyObject *_wrap_device_name_get(PyObject *SWIGUNUSEDPARM(self), PyObj
     SWIG_exception_fail(SWIG_ArgError(res1), "in method '" "device_name_get" "', argument " "1"" of type '" "device *""'"); 
   }
   arg1 = reinterpret_cast< device * >(argp1);
-  result = (std::string *) & ((arg1)->name);
+  {
+    SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+    result = (std::string *) & ((arg1)->name);
+    SWIG_PYTHON_THREAD_END_ALLOW;
+  }
   resultobj = SWIG_From_std_string(static_cast< std::string >(*result));
   return resultobj;
 fail:
@@ -8859,7 +9184,11 @@ SWIGINTERN PyObject *_wrap_device_lastSeen_set(PyObject *SWIGUNUSEDPARM(self), P
       if (SWIG_IsNewObj(res2)) delete temp;
     }
   }
-  if (arg1) (arg1)->lastSeen = arg2;
+  {
+    SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+    if (arg1) (arg1)->lastSeen = arg2;
+    SWIG_PYTHON_THREAD_END_ALLOW;
+  }
   resultobj = SWIG_Py_Void();
   return resultobj;
 fail:
@@ -8882,7 +9211,11 @@ SWIGINTERN PyObject *_wrap_device_lastSeen_get(PyObject *SWIGUNUSEDPARM(self), P
     SWIG_exception_fail(SWIG_ArgError(res1), "in method '" "device_lastSeen_get" "', argument " "1"" of type '" "device *""'"); 
   }
   arg1 = reinterpret_cast< device * >(argp1);
-  result =  ((arg1)->lastSeen);
+  {
+    SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+    result =  ((arg1)->lastSeen);
+    SWIG_PYTHON_THREAD_END_ALLOW;
+  }
   resultobj = SWIG_NewPointerObj((new std::time_t(static_cast< const std::time_t& >(result))), SWIGTYPE_p_std__time_t, SWIG_POINTER_OWN |  0 );
   return resultobj;
 fail:
@@ -8919,7 +9252,11 @@ SWIGINTERN PyObject *_wrap_device_lastUsed_set(PyObject *SWIGUNUSEDPARM(self), P
       if (SWIG_IsNewObj(res2)) delete temp;
     }
   }
-  if (arg1) (arg1)->lastUsed = arg2;
+  {
+    SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+    if (arg1) (arg1)->lastUsed = arg2;
+    SWIG_PYTHON_THREAD_END_ALLOW;
+  }
   resultobj = SWIG_Py_Void();
   return resultobj;
 fail:
@@ -8942,7 +9279,11 @@ SWIGINTERN PyObject *_wrap_device_lastUsed_get(PyObject *SWIGUNUSEDPARM(self), P
     SWIG_exception_fail(SWIG_ArgError(res1), "in method '" "device_lastUsed_get" "', argument " "1"" of type '" "device *""'"); 
   }
   arg1 = reinterpret_cast< device * >(argp1);
-  result =  ((arg1)->lastUsed);
+  {
+    SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+    result =  ((arg1)->lastUsed);
+    SWIG_PYTHON_THREAD_END_ALLOW;
+  }
   resultobj = SWIG_NewPointerObj((new std::time_t(static_cast< const std::time_t& >(result))), SWIGTYPE_p_std__time_t, SWIG_POINTER_OWN |  0 );
   return resultobj;
 fail:
@@ -8971,7 +9312,11 @@ SWIGINTERN PyObject *_wrap_device_connected_set(PyObject *SWIGUNUSEDPARM(self), 
     SWIG_exception_fail(SWIG_ArgError(ecode2), "in method '" "device_connected_set" "', argument " "2"" of type '" "bool""'");
   } 
   arg2 = static_cast< bool >(val2);
-  if (arg1) (arg1)->connected = arg2;
+  {
+    SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+    if (arg1) (arg1)->connected = arg2;
+    SWIG_PYTHON_THREAD_END_ALLOW;
+  }
   resultobj = SWIG_Py_Void();
   return resultobj;
 fail:
@@ -8994,7 +9339,11 @@ SWIGINTERN PyObject *_wrap_device_connected_get(PyObject *SWIGUNUSEDPARM(self), 
     SWIG_exception_fail(SWIG_ArgError(res1), "in method '" "device_connected_get" "', argument " "1"" of type '" "device *""'"); 
   }
   arg1 = reinterpret_cast< device * >(argp1);
-  result = (bool) ((arg1)->connected);
+  {
+    SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+    result = (bool) ((arg1)->connected);
+    SWIG_PYTHON_THREAD_END_ALLOW;
+  }
   resultobj = SWIG_From_bool(static_cast< bool >(result));
   return resultobj;
 fail:
@@ -9023,7 +9372,11 @@ SWIGINTERN PyObject *_wrap_device_remembered_set(PyObject *SWIGUNUSEDPARM(self),
     SWIG_exception_fail(SWIG_ArgError(ecode2), "in method '" "device_remembered_set" "', argument " "2"" of type '" "bool""'");
   } 
   arg2 = static_cast< bool >(val2);
-  if (arg1) (arg1)->remembered = arg2;
+  {
+    SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+    if (arg1) (arg1)->remembered = arg2;
+    SWIG_PYTHON_THREAD_END_ALLOW;
+  }
   resultobj = SWIG_Py_Void();
   return resultobj;
 fail:
@@ -9046,7 +9399,11 @@ SWIGINTERN PyObject *_wrap_device_remembered_get(PyObject *SWIGUNUSEDPARM(self),
     SWIG_exception_fail(SWIG_ArgError(res1), "in method '" "device_remembered_get" "', argument " "1"" of type '" "device *""'"); 
   }
   arg1 = reinterpret_cast< device * >(argp1);
-  result = (bool) ((arg1)->remembered);
+  {
+    SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+    result = (bool) ((arg1)->remembered);
+    SWIG_PYTHON_THREAD_END_ALLOW;
+  }
   resultobj = SWIG_From_bool(static_cast< bool >(result));
   return resultobj;
 fail:
@@ -9075,7 +9432,11 @@ SWIGINTERN PyObject *_wrap_device_authenticated_set(PyObject *SWIGUNUSEDPARM(sel
     SWIG_exception_fail(SWIG_ArgError(ecode2), "in method '" "device_authenticated_set" "', argument " "2"" of type '" "bool""'");
   } 
   arg2 = static_cast< bool >(val2);
-  if (arg1) (arg1)->authenticated = arg2;
+  {
+    SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+    if (arg1) (arg1)->authenticated = arg2;
+    SWIG_PYTHON_THREAD_END_ALLOW;
+  }
   resultobj = SWIG_Py_Void();
   return resultobj;
 fail:
@@ -9098,7 +9459,11 @@ SWIGINTERN PyObject *_wrap_device_authenticated_get(PyObject *SWIGUNUSEDPARM(sel
     SWIG_exception_fail(SWIG_ArgError(res1), "in method '" "device_authenticated_get" "', argument " "1"" of type '" "device *""'"); 
   }
   arg1 = reinterpret_cast< device * >(argp1);
-  result = (bool) ((arg1)->authenticated);
+  {
+    SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+    result = (bool) ((arg1)->authenticated);
+    SWIG_PYTHON_THREAD_END_ALLOW;
+  }
   resultobj = SWIG_From_bool(static_cast< bool >(result));
   return resultobj;
 fail:
@@ -9113,7 +9478,11 @@ SWIGINTERN PyObject *_wrap_new_device(PyObject *SWIGUNUSEDPARM(self), PyObject *
   if (!SWIG_Python_UnpackTuple(args, "new_device", 0, 0, 0)) SWIG_fail;
   {
     try {
-      result = (device *)new device();
+      {
+        SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+        result = (device *)new device();
+        SWIG_PYTHON_THREAD_END_ALLOW;
+      }
     } catch(const ExploreException& e) {
       SWIG_exception(SWIG_ValueError, e.what());
     }
@@ -9149,7 +9518,11 @@ SWIGINTERN PyObject *_wrap_delete_device(PyObject *SWIGUNUSEDPARM(self), PyObjec
   arg1 = reinterpret_cast< device * >(argp1);
   {
     try {
-      delete arg1;
+      {
+        SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+        delete arg1;
+        SWIG_PYTHON_THREAD_END_ALLOW;
+      }
     } catch(const ExploreException& e) {
       SWIG_exception(SWIG_ValueError, e.what());
     }
@@ -9196,7 +9569,11 @@ SWIGINTERN PyObject *_wrap_delete_ExploreSDK(PyObject *SWIGUNUSEDPARM(self), PyO
   arg1 = reinterpret_cast< ExploreSDK * >(argp1);
   {
     try {
-      delete arg1;
+      {
+        SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+        delete arg1;
+        SWIG_PYTHON_THREAD_END_ALLOW;
+      }
     } catch(const ExploreException& e) {
       SWIG_exception(SWIG_ValueError, e.what());
     }
@@ -9223,7 +9600,11 @@ SWIGINTERN PyObject *_wrap_ExploreSDK_Create(PyObject *SWIGUNUSEDPARM(self), PyO
   if (!SWIG_Python_UnpackTuple(args, "ExploreSDK_Create", 0, 0, 0)) SWIG_fail;
   {
     try {
-      result = (ExploreSDK *)ExploreSDK::Create();
+      {
+        SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+        result = (ExploreSDK *)ExploreSDK::Create();
+        SWIG_PYTHON_THREAD_END_ALLOW;
+      }
     } catch(const ExploreException& e) {
       SWIG_exception(SWIG_ValueError, e.what());
     }
@@ -9266,7 +9647,11 @@ SWIGINTERN PyObject *_wrap_ExploreSDK_PerformDeviceSearch__SWIG_0(PyObject *SWIG
   arg2 = static_cast< int >(val2);
   {
     try {
-      result = (arg1)->PerformDeviceSearch(arg2);
+      {
+        SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+        result = (arg1)->PerformDeviceSearch(arg2);
+        SWIG_PYTHON_THREAD_END_ALLOW;
+      }
     } catch(const ExploreException& e) {
       SWIG_exception(SWIG_ValueError, e.what());
     }
@@ -9301,7 +9686,11 @@ SWIGINTERN PyObject *_wrap_ExploreSDK_PerformDeviceSearch__SWIG_1(PyObject *SWIG
   arg1 = reinterpret_cast< ExploreSDK * >(argp1);
   {
     try {
-      result = (arg1)->PerformDeviceSearch();
+      {
+        SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+        result = (arg1)->PerformDeviceSearch();
+        SWIG_PYTHON_THREAD_END_ALLOW;
+      }
     } catch(const ExploreException& e) {
       SWIG_exception(SWIG_ValueError, e.what());
     }
@@ -9389,7 +9778,11 @@ SWIGINTERN PyObject *_wrap_ExploreSDK_SdpSearch(PyObject *SWIGUNUSEDPARM(self), 
   }
   {
     try {
-      result = (int)(arg1)->SdpSearch(arg2);
+      {
+        SWIG_PYTHON_THREAD_BEGIN_ALLOW;
+        result = (int)(arg1)->SdpSearch(arg2);
+        SWIG_PYTHON_THREAD_END_ALLOW;
+      }
     } catch(const ExploreException& e) {
       SWIG_exception(SWIG_ValueError, e.what());
     }
@@ -10391,6 +10784,9 @@ SWIG_init(void) {
   // thread safe initialization
   swig::container_owner_attribute();
   
+  
+  /* Initialize threading */
+  SWIG_PYTHON_INITIALIZE_THREADS;
 #if PY_VERSION_HEX >= 0x03000000
   return m;
 #else